decomposition of charts component + folder organisation
